+++ app/controllers/usuarios_controller.rb
@@ -109,42 +109,6 @@ class UsuariosController < ApplicationController
         end
     end
 
-    def get_usuarios
-		respond_to do |format|
-			format.json { render json: { usuarios: Usuario.all }, status: :ok }
-		end
-    end
-
-    def create_usuario
-        json_params = JSON.parse(request.raw_post)
-        novo_usuario = Usuario.new
-        novo_usuario.nome_completo = json_params['nome']
-        novo_usuario.pronomes_usuarios_id = json_params['pronomes']
-        novo_usuario.email = json_params['email']
-        novo_usuario.password = json_params['senha']
-        novo_usuario.usuario_cargo_id = json_params['cargo']
-        novo_usuario.acessos_count = 0
-        
-        if novo_usuario.save then
-            if novo_usuario.pronomes_usuarios_id == 1 then
-                logtxt = "[EXT] Usuária " + novo_usuario.nome_completo + " cadastrada com sucesso."
-            else
-                logtxt = "[EXT] Usuário " + novo_usuario.nome_completo + " cadastrado com sucesso."
-            end
-            respond_to do |format|
-                format.json { render json: novo_usuario, status: :ok }
-            end
-        else
-            Rails.logger.error "[EXT] Houve um erro ao cadastrar o usuário."
-            respond_to do |format|
-                format.json { render json: {
-                    json: json_params,
-                    erros: novo_usuario.errors
-                }, status: :bad_request }
-            end
-        end
-    end
-
     private
 
         def usuario_params
