+++ app/helpers/usuarios_ext_helper.rb
@@ -1,0 +1,2 @@
+module UsuariosExtHelper
+end
